move permissions from lms to users
